fix(script.js) split script.js file into smaller pieces using export and inport
