+++ index.html
@@ -276,5 +276,5 @@
     </div>
 </footer>
   </body>
-  <script src="js/script.js"></script>
+  <script type="module" src="js/script.js"></script>
 </html>
