+++ js/main/slyder.js
@@ -1,0 +1,8 @@
+let counter = 1;
+export const interval = setInterval(function () {
+  document.getElementById("radio" + counter).checked = true;
+  counter++;
+  if (counter > 3) {
+    counter = 1;
+  }
+}, 2000);
